refactor tests: moved tests to proper packages, renaming
