+++ src/test/scala/com/redislabs/provider/redis/df/cluster/DataframeClusterSuite.scala
@@ -1,6 +1,6 @@
 package com.redislabs.provider.redis.df.cluster
 
 import com.redislabs.provider.redis.df.DataframeSuite
-import com.redislabs.provider.redis.rdd.RedisClusterEnv
+import com.redislabs.provider.redis.env.RedisClusterEnv
 
 class DataframeClusterSuite extends DataframeSuite with RedisClusterEnv
