+++ src/test/scala/com/redislabs/provider/redis/df/cluster/SparkSqlClusterSuite.scala
@@ -1,7 +1,7 @@
 package com.redislabs.provider.redis.df.cluster
 
 import com.redislabs.provider.redis.df.SparkSqlSuite
-import com.redislabs.provider.redis.rdd.RedisClusterEnv
+import com.redislabs.provider.redis.env.RedisClusterEnv
 
 /**
   * @author The Viet Nguyen
