+++ src/test/scala/com/redislabs/provider/redis/df/standalone/DataframeStandaloneSuite.scala
@@ -1,6 +1,6 @@
 package com.redislabs.provider.redis.df.standalone
 
 import com.redislabs.provider.redis.df.DataframeSuite
-import com.redislabs.provider.redis.rdd.RedisStandaloneEnv
+import com.redislabs.provider.redis.env.RedisStandaloneEnv
 
 class DataframeStandaloneSuite extends DataframeSuite with RedisStandaloneEnv
