+++ src/test/scala/com/redislabs/provider/redis/df/standalone/FilteredDataframeStandaloneSuite.scala
@@ -1,7 +1,7 @@
 package com.redislabs.provider.redis.df.standalone
 
 import com.redislabs.provider.redis.df.RedisDataframeSuite
-import com.redislabs.provider.redis.rdd.RedisStandaloneEnv
+import com.redislabs.provider.redis.env.RedisStandaloneEnv
 
 /**
   * @author The Viet Nguyen
