+++ src/test/scala/com/redislabs/provider/redis/rdd/cluster/RedisKeysClusterSuite.scala
@@ -1,5 +1,6 @@
 package com.redislabs.provider.redis.rdd.cluster
 
-import com.redislabs.provider.redis.rdd.{RedisClusterEnv, RedisKeysSuite}
+import com.redislabs.provider.redis.env.RedisClusterEnv
+import com.redislabs.provider.redis.rdd.RedisKeysSuite
 
 class RedisKeysClusterSuite extends RedisKeysSuite with RedisClusterEnv
