+++ src/test/scala/com/redislabs/provider/redis/rdd/cluster/RedisRDDClusterSuite.scala
@@ -1,5 +1,6 @@
 package com.redislabs.provider.redis.rdd.cluster
 
-import com.redislabs.provider.redis.rdd.{RedisClusterEnv, RedisRddSuite}
+import com.redislabs.provider.redis.env.RedisClusterEnv
+import com.redislabs.provider.redis.rdd.RedisRddSuite
 
 class RedisRDDClusterSuite extends RedisRddSuite with RedisClusterEnv
