+++ src/test/scala/com/redislabs/provider/redis/rdd/standalone/RedisKeysStandaloneSuite.scala
@@ -1,5 +1,6 @@
 package com.redislabs.provider.redis.rdd.standalone
 
-import com.redislabs.provider.redis.rdd.{RedisKeysSuite, RedisStandaloneEnv}
+import com.redislabs.provider.redis.env.RedisStandaloneEnv
+import com.redislabs.provider.redis.rdd.RedisKeysSuite
 
 class RedisKeysStandaloneSuite extends RedisKeysSuite with RedisStandaloneEnv
