+++ src/test/scala/com/redislabs/provider/redis/rdd/standalone/RedisRDDStandaloneSuite.scala
@@ -1,5 +1,6 @@
 package com.redislabs.provider.redis.rdd.standalone
 
-import com.redislabs.provider.redis.rdd.{RedisRddSuite, RedisStandaloneEnv}
+import com.redislabs.provider.redis.env.RedisStandaloneEnv
+import com.redislabs.provider.redis.rdd.RedisRddSuite
 
 class RedisRDDStandaloneSuite extends RedisRddSuite with RedisStandaloneEnv
